+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1
+### Changed
+- Remove the localeAlternate() Twig function, use the default getSwitcherSites with default params instead.
+- Rename the langSwitcher() Twig function to getSwitcherSites(). Note that the legacy name is still supported.
+
 ## 1.0.4
 ### Changed
 - Changed issues and changelog URLs